Code/app-backend/requirements.txt
amqp==5.1.1
async-timeout==4.0.2
bcrypt==4.0.1
billiard==4.1.0
blinker==1.6.2
Brotli==1.0.9
cachelib==0.9.0
celery==5.3.1
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.1
cssselect2==0.7.0
ecdsa==0.18.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.2
Flask-Celery-Helper==1.1.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-redis==0.4.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
flower==2.0.0
fonttools==4.42.0
greenlet==2.0.2
html5lib==1.1
humanize==4.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.3
pdfkit==1.0.0
Pillow==10.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
pyasn1==0.5.0
pycparser==2.21
pydyf==0.7.0
PyJWT
pyphen==0.14.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2023.3
qrcode==7.4.2
redis==4.6.0
reportlab==4.0.4
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.17
tinycss2==1.2.1
tornado==6.3.2
typing_extensions==4.6.3
tzdata==2023.3
uuid==1.30
vine==5.0.0
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
Werkzeug==2.3.6
WTForms==3.0.1
zopfli==0.2.2
